feat:transaction crud and validation of dto

+++ ms-transactions/prisma/migrations/20231221135623_init/migration.sql
@@ -5,7 +5,7 @@ CREATE TYPE "Operation" AS ENUM ('CREDIT', 'DEBIT');
 CREATE TABLE "Transaction" (
     "id" SERIAL NOT NULL,
     "user_id" TEXT NOT NULL,
-    "amount" BIGINT NOT NULL,
+    "amount" INTEGER NOT NULL,
     "type" "Operation" NOT NULL,
 
     CONSTRAINT "Transaction_pkey" PRIMARY KEY ("id")
